+++ packages/pupperfish-react/src/index.ts
@@ -3,4 +3,5 @@ export * from "./pupperfishSignal.js";
 export * from "./PupperfishDock.js";
 export * from "./PupperfishWidgetShell.js";
 export * from "./PupperfishChatShell.js";
+export * from "./PupperfishChartViewer.js";
 export * from "./TradeImageGalleryManager.js";